requirements.txt
Flask
line-bot-sdk
requests
gunicorn
google-generativeai